feat(common): add physical, virtual, and IPA address types with arithmetic operations and MMIO region struct
