+++ arch_hal/aarch64_hal/Cargo.toml
@@ -21,3 +21,4 @@ aarch64_mutex = { path = "./aarch64_mutex" }
 aarch64_gdb = { path = "./aarch64_gdb" }
 soc = { path = "./soc" }
 tls = { path = "./tls" }
+common = { path = "./common" }
