+++ arch_hal/aarch64_hal/common/src/addr.rs
@@ -1,0 +1,101 @@
+#[repr(transparent)]
+#[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
+pub struct PhysAddr(usize);
+
+impl PhysAddr {
+    pub const fn new(value: usize) -> Self {
+        Self(value)
+    }
+
+    pub const fn as_usize(self) -> usize {
+        self.0
+    }
+
+    pub const fn checked_add(self, rhs: usize) -> Option<Self> {
+        match self.0.checked_add(rhs) {
+            Some(v) => Some(Self(v)),
+            None => None,
+        }
+    }
+
+    pub const fn checked_sub(self, rhs: usize) -> Option<Self> {
+        match self.0.checked_sub(rhs) {
+            Some(v) => Some(Self(v)),
+            None => None,
+        }
+    }
+}
+
+#[repr(transparent)]
+#[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
+pub struct VirtAddr(usize);
+
+impl VirtAddr {
+    pub const fn new(value: usize) -> Self {
+        Self(value)
+    }
+
+    pub const fn as_usize(self) -> usize {
+        self.0
+    }
+
+    pub const fn checked_add(self, rhs: usize) -> Option<Self> {
+        match self.0.checked_add(rhs) {
+            Some(v) => Some(Self(v)),
+            None => None,
+        }
+    }
+
+    pub const fn checked_sub(self, rhs: usize) -> Option<Self> {
+        match self.0.checked_sub(rhs) {
+            Some(v) => Some(Self(v)),
+            None => None,
+        }
+    }
+}
+
+#[repr(transparent)]
+#[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
+pub struct IpaAddr(usize);
+
+impl IpaAddr {
+    pub const fn new(value: usize) -> Self {
+        Self(value)
+    }
+
+    pub const fn as_usize(self) -> usize {
+        self.0
+    }
+
+    pub const fn checked_add(self, rhs: usize) -> Option<Self> {
+        match self.0.checked_add(rhs) {
+            Some(v) => Some(Self(v)),
+            None => None,
+        }
+    }
+
+    pub const fn checked_sub(self, rhs: usize) -> Option<Self> {
+        match self.0.checked_sub(rhs) {
+            Some(v) => Some(Self(v)),
+            None => None,
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn checked_arithmetic() {
+        let p = PhysAddr::new(0x1000);
+        assert_eq!(p.checked_add(0x20).unwrap().as_usize(), 0x1020);
+        assert_eq!(p.checked_sub(0x20).unwrap().as_usize(), 0x0fe0);
+        assert!(PhysAddr::new(usize::MAX).checked_add(1).is_none());
+        assert!(VirtAddr::new(0).checked_sub(1).is_none());
+        assert_eq!(
+            IpaAddr::new(0x2000).checked_add(0x10).unwrap().as_usize(),
+            0x2010
+        );
+    }
+}
