+++ arch_hal/aarch64_hal/exceptions/Cargo.toml
@@ -12,6 +12,7 @@ typestate = { path = "../../../typestate" }
 typestate_macro = { path = "../../../typestate_macro" }
 print = { path = "../print" }
 psci = { path = "../psci" }
+common = { path = "../common" }
 
 [features]
 testapi = []
