+++ arch_hal/aarch64_hal/gic/src/lib.rs
@@ -34,6 +34,7 @@ pub mod dt_irq;
 pub mod vm;
 
 pub use common::IrqSense;
+pub use common::MmioRegion;
 pub use common::TriggerMode;
 use cpu::CoreAffinity;
 
@@ -41,21 +42,6 @@ use cpu::CoreAffinity;
 use exceptions::setup_exception;
 #[cfg(all(test, target_arch = "aarch64"))]
 extern crate alloc;
-
-/// An MMIO region describing a device register frame.
-///
-/// `base` is the physical (or already-mapped virtual) base address used for volatile access.
-/// `size` is the byte size of the region.
-///
-/// Implementations typically require:
-/// - `base` aligned to the frame granule (commonly 4KiB),
-/// - a mapping with *Device* memory attributes (no speculative access, strongly ordered as required),
-/// - a region size matching the register block layout used by the backend.
-#[derive(Copy, Clone, Debug)]
-pub struct MmioRegion {
-    pub base: usize,
-    pub size: usize,
-}
 
 /// Error type used by all GIC traits.
 ///
