+++ xtest.txt
@@ -7,6 +7,7 @@
 #   uboot       <package> <testname> <testscript> [extra...]
 
 std allocator
+std common
 std dtb
 std intrusive_linked_list
 std mutex
